+++ CHANGES.md
@@ -19,9 +19,15 @@
   - @torikizi
 - [UPDATE] Intel VPL を v2.16.0 にあげる
   - @torikizi
+- [ADD] Android x86_64 ターゲットを追加する
+  - 既存の android ターゲットとは別で x86_64 専用のターゲットを用意する
+  - @zztkm
 
 ### misc
 
+- [CHANGE] GitHub Actions から Android x86_64 以外のターゲット向けのビルドを削除する
+  - support/android-x86_64 ブランチでは Android x86_64 以外のターゲットをビルドする必要がないため削除する
+  - @zztkm
 - [UPDATE] Examples の DEPS を更新する
   - WEBRTC_BUILD_VERSION を m144.7559.0.0 にあげる
   - CMake を 4.2.1 にあげる